Decrease g++ compilation memory consumption
