+++ test/core/stream/CMakeLists.txt
@@ -6,12 +6,18 @@ add_boost_test("test.core.stream.create" "create.cpp" "")
 
 add_boost_test("test.core.stream.drop" "drop.cpp" "")
 
-add_boost_test("test.core.stream.filter" "filter.cpp" "")
+add_boost_test("test.core.stream.filter1" "filter1.cpp" "")
+
+add_boost_test("test.core.stream.filter2" "filter2.cpp" "")
 
 add_boost_test("test.core.stream.for_each" "for_each.cpp" "")
 
 add_boost_test("test.core.stream.random_fill" "random_fill.cpp" "")
 
-add_boost_test("test.core.stream.transform" "transform.cpp" "")
+add_boost_test("test.core.stream.transform1" "transform1.cpp" "")
+
+add_boost_test("test.core.stream.transform2" "transform2.cpp" "")
+
+add_boost_test("test.core.stream.transform3" "transform3.cpp" "")
 
 add_boost_test("test.core.stream.unpack" "unpack.cpp" "")
