+++ test/core/stream/filter.hpp
@@ -1,0 +1,26 @@
+/****************************************************************************
+ *  cxtream library
+ *  Copyright (c) 2017, Cognexa Solutions s.r.o.
+ *  Author(s) Filip Matzner
+ *
+ *  This file is distributed under the MIT License.
+ *  See the accompanying file LICENSE.txt for the complete license agreement.
+ ****************************************************************************/
+
+#ifndef TEST_STREAM_FILTER_HPP
+#define TEST_STREAM_FILTER_HPP
+
+#include "../common.hpp"
+
+#include <cxtream/core/stream/create.hpp>
+#include <cxtream/core/stream/filter.hpp>
+#include <cxtream/core/stream/for_each.hpp>
+
+#include <boost/test/unit_test.hpp>
+#include <range/v3/to_container.hpp>
+#include <range/v3/view/move.hpp>
+
+#include <tuple>
+#include <vector>
+
+#endif
